tests - refactor, separated lane tests
