+++ test/component/parser/json/BpmnJsonParser.test.ts
@@ -1,33 +1,7 @@
 import { expect } from 'chai';
 import BpmnJsonParser from '../../../../src/component/parser/json/BpmnJsonParser';
 import { ShapeBpmnElementKind } from '../../../../src/model/bpmn/shape/ShapeBpmnElementKind';
-import Shape from '../../../../src/model/bpmn/shape/Shape';
-
-interface ExpectedShape {
-  shapeId: string;
-  bpmnElementId: string;
-  bpmnElementName: string;
-  bpmnElementKind: ShapeBpmnElementKind;
-  boundsX: number;
-  boundsY: number;
-  boundsWidth: number;
-  boundsHeight: number;
-}
-
-function verifyShape(shape: Shape, expectedValue: ExpectedShape) {
-  expect(shape.id).to.be.equal(expectedValue.shapeId, 'shape id');
-
-  const bpmnElement = shape.bpmnElement;
-  expect(bpmnElement.id).to.be.equal(expectedValue.bpmnElementId, 'bpmn element id');
-  expect(bpmnElement.name).to.be.equal(expectedValue.bpmnElementName, 'bpmn element name');
-  expect(bpmnElement.kind).to.be.equal(expectedValue.bpmnElementKind, 'bpmn element kind');
-
-  const bounds = shape.bounds;
-  expect(bounds.x).to.be.equal(expectedValue.boundsX, 'bounds x');
-  expect(bounds.y).to.be.equal(expectedValue.boundsY, 'bounds y');
-  expect(bounds.width).to.be.equal(expectedValue.boundsWidth, 'bounds width');
-  expect(bounds.height).to.be.equal(expectedValue.boundsHeight, 'bounds height');
-}
+import { verifyShape } from './util';
 
 describe('parse bpmn as json', () => {
   describe('start event', () => {
@@ -433,239 +407,6 @@ describe('parse bpmn as json', () => {
       expect(bpmnElementWithoutName.id).to.be.equal('sequenceFlow_id_1', 'bpmn element id');
       expect(bpmnElementWithoutName.sourceRefId).to.be.equal('sequenceFlow_id_1', 'bpmn element sourceRef');
       expect(bpmnElementWithoutName.targetRefId).to.be.equal('targetRef_1', 'bpmn element targetRef');
-    });
-  });
-
-  describe('lane', () => {
-    it('json containing one process with a single lane without flowNodeRef', () => {
-      const json = `{
-                      "definitions":{
-                        "process":{
-                          "lane": { "id":"Lane_12u5n6x" }
-                        },
-                        "BPMNDiagram":{
-                          "BPMNPlane":{
-                            "BPMNShape":
-                              {
-                                "id":"Lane_1h5yeu4_di",
-                                "bpmnElement":"Lane_12u5n6x",
-                                "isHorizontal":true,
-                                "Bounds": { "x":362, "y":232, "width":36, "height":45 }
-                              }
-                          }
-                        }
-                      }
-                    }`;
-
-      const model = BpmnJsonParser.parse(JSON.parse(json));
-
-      expect(model.lanes).to.have.lengthOf(1, 'lanes');
-      verifyShape(model.lanes[0], {
-        shapeId: 'Lane_1h5yeu4_di',
-        bpmnElementId: 'Lane_12u5n6x',
-        bpmnElementName: undefined,
-        bpmnElementKind: ShapeBpmnElementKind.LANE,
-        boundsX: 362,
-        boundsY: 232,
-        boundsWidth: 36,
-        boundsHeight: 45,
-      });
-
-      expect(model.flowNodes).to.have.lengthOf(0, 'shapes');
-    });
-
-    it('json containing one process with a single lane with flowNodeRef as object', () => {
-      const json = `{
-                      "definitions":{
-                        "process":{
-                          "lane": { "id":"Lane_12u5n6x", "flowNodeRef":"event_id_0" },
-                          "startEvent": { "id":"event_id_0" }
-                        },
-                        "BPMNDiagram":{
-                          "BPMNPlane":{
-                            "BPMNShape":[
-                              {
-                                "id":"Lane_1h5yeu4_di",
-                                "bpmnElement":"Lane_12u5n6x",
-                                "isHorizontal":true,
-                                "Bounds": { "x":362, "y":232, "width":36, "height":45 }
-                              },
-                              {
-                                "id":"event_id_0_di",
-                                "bpmnElement":"event_id_0",
-                                "Bounds":{ "x":11, "y":11, "width":11, "height":11 }
-                              }
-                            ]
-                          }
-                        }
-                      }
-                    }`;
-
-      const model = BpmnJsonParser.parse(JSON.parse(json));
-
-      expect(model.lanes).to.have.lengthOf(1, 'lanes');
-      verifyShape(model.lanes[0], {
-        shapeId: 'Lane_1h5yeu4_di',
-        bpmnElementId: 'Lane_12u5n6x',
-        bpmnElementName: undefined,
-        bpmnElementKind: ShapeBpmnElementKind.LANE,
-        boundsX: 362,
-        boundsY: 232,
-        boundsWidth: 36,
-        boundsHeight: 45,
-      });
-
-      expect(model.flowNodes).to.have.lengthOf(1, 'shapes');
-      expect(model.flowNodes[0].bpmnElement.parentId).to.be.equal('Lane_12u5n6x', 'bpmn element parent');
-    });
-
-    it('json containing one process with a single lane with flowNodeRef as array', () => {
-      const json = `{
-                      "definitions":{
-                        "process":{
-                          "lane": { "id":"Lane_12u5n6x", "flowNodeRef":["event_id_0"] },
-                          "startEvent": { "id":"event_id_0" }
-                        },
-                        "BPMNDiagram":{
-                          "BPMNPlane":{
-                            "BPMNShape":[
-                              {
-                                "id":"Lane_1h5yeu4_di",
-                                "bpmnElement":"Lane_12u5n6x",
-                                "isHorizontal":true,
-                                "Bounds": { "x":362, "y":232, "width":36, "height":45 }
-                              },
-                              {
-                                "id":"event_id_0_di",
-                                "bpmnElement":"event_id_0",
-                                "Bounds":{ "x":11, "y":11, "width":11, "height":11 }
-                              }
-                            ]
-                          }
-                        }
-                      }
-                    }`;
-
-      const model = BpmnJsonParser.parse(JSON.parse(json));
-
-      expect(model.lanes).to.have.lengthOf(1, 'lanes');
-      verifyShape(model.lanes[0], {
-        shapeId: 'Lane_1h5yeu4_di',
-        bpmnElementId: 'Lane_12u5n6x',
-        bpmnElementName: undefined,
-        bpmnElementKind: ShapeBpmnElementKind.LANE,
-        boundsX: 362,
-        boundsY: 232,
-        boundsWidth: 36,
-        boundsHeight: 45,
-      });
-
-      expect(model.flowNodes).to.have.lengthOf(1, 'shapes');
-      expect(model.flowNodes[0].bpmnElement.parentId).to.be.equal('Lane_12u5n6x', 'bpmn element parent');
-    });
-
-    it('json containing one process declared as array with a laneset', () => {
-      const json = `{
-                      "definitions":{
-                        "process":[{
-                          "laneSet":{
-                            "id":"LaneSet_1i59xiy",
-                            "lane":{ "id":"Lane_12u5n6x" }
-                          }
-                        }],
-                        "BPMNDiagram":{
-                          "BPMNPlane":{
-                            "BPMNShape":
-                              {
-                                "id":"Lane_1h5yeu4_di",
-                                "bpmnElement":"Lane_12u5n6x",
-                                "isHorizontal":true,
-                                "Bounds":{ "x":362, "y":232, "width":36, "height":45 }
-                              }
-                          }
-                        }
-                      }
-                    }`;
-
-      const model = BpmnJsonParser.parse(JSON.parse(json));
-      expect(model.flowNodes).to.have.lengthOf(0, 'shapes');
-
-      expect(model.lanes).to.have.lengthOf(1, 'lanes');
-      verifyShape(model.lanes[0], {
-        shapeId: 'Lane_1h5yeu4_di',
-        bpmnElementId: 'Lane_12u5n6x',
-        bpmnElementName: undefined,
-        bpmnElementKind: ShapeBpmnElementKind.LANE,
-        boundsX: 362,
-        boundsY: 232,
-        boundsWidth: 36,
-        boundsHeight: 45,
-      });
-    });
-
-    it('json containing one process with an array of lanes with & without name', () => {
-      const json = `{
-                      "definitions":{
-                        "process":{
-                          "laneSet":{
-                            "id":"LaneSet_1i59xiy",
-                            "lane":[
-                              {
-                                "id":"Lane_164yevk",
-                                "name":"Customer",
-                                "flowNodeRef":"event_id_0"
-                              },
-                              { "id":"Lane_12u5n6x" }
-                            ]
-                          },
-                          "startEvent": { "id":"event_id_0" }
-                        },
-                        "BPMNDiagram":{
-                          "BPMNPlane":{
-                            "BPMNShape":[
-                              {
-                                "id":"Lane_164yevk_di",
-                                "bpmnElement":"Lane_164yevk",
-                                "isHorizontal":true,
-                                "Bounds":{ "x":362, "y":232, "width":36, "height":45 }
-                              },
-                              {
-                                "id":"Lane_12u5n6x_di",
-                                "bpmnElement":"Lane_12u5n6x",
-                                "isHorizontal":true,
-                                "Bounds":{ "x":666, "y":222, "width":22, "height":33 }
-                              }
-                            ]
-                          }
-                        }
-                      }
-                    }`;
-
-      const model = BpmnJsonParser.parse(JSON.parse(json));
-
-      expect(model.flowNodes).to.have.lengthOf(0, 'shapes');
-
-      expect(model.lanes).to.have.lengthOf(2, 'lanes');
-      verifyShape(model.lanes[0], {
-        shapeId: 'Lane_164yevk_di',
-        bpmnElementId: 'Lane_164yevk',
-        bpmnElementName: 'Customer',
-        bpmnElementKind: ShapeBpmnElementKind.LANE,
-        boundsX: 362,
-        boundsY: 232,
-        boundsWidth: 36,
-        boundsHeight: 45,
-      });
-      verifyShape(model.lanes[1], {
-        shapeId: 'Lane_12u5n6x_di',
-        bpmnElementId: 'Lane_12u5n6x',
-        bpmnElementName: undefined,
-        bpmnElementKind: ShapeBpmnElementKind.LANE,
-        boundsX: 666,
-        boundsY: 222,
-        boundsWidth: 22,
-        boundsHeight: 33,
-      });
     });
   });
 });
